move octokit to its own module
introduce isGithubOrganizationMember
